refactor: Rebased project to Thymeleaf
