+++ .instructions/2. core exercises.md
@@ -38,14 +38,14 @@ The following exercises will help you get started with GitHub Copilot. You must 
 
 ### Adding the buttons to the calculator UI
 
-1. Open the ```static/index.html``` file in the editor window.
+1. Open the ```templates/index.html``` file in the editor window.
 
-2. Scroll down to where you see the ```<!-- TODO: Buttons -->``` comment
+2. Scroll down to where you see the ```<!-- TODO #1: Buttons -->``` comment
 
 3. Add a new line below this comment and type the following two lines. You should see GitHub Copilot start to autocomplete the second line as you type. When you see this, just press ```TAB``` to accept the completion.
 
 ``` <!-- add a button for a power (or exponential) function --> ```
-``` <button class="btn" onClick="operationPressed('^')">^</button> ```
+
 
 Your finished snippet should match the following.
 
@@ -53,42 +53,42 @@ Your finished snippet should match the following.
 
 ### Adding the logic for the new features
 
-5. Open the 
-[ArithmeticController.java](../src/main/java/com/copilot/calculator/api/ArithmeticController.java#L16) file in the editor window.
-
-6. Scroll down to where you see the ```// TODO: Add operator``` comment
-
-7. Press **ENTER** at the end of the line that defines the divide function.
-
-8. Start typing the following line and notice that GitHub Copilot should start to offer code completion half way through the word "power" as you're typing. Press **TAB** to accept the suggestion.
-
-```"power",    (Double a, Double b) -> Math.pow(a, b)```
-
-9. Open the 
+4. Open the 
 [client.js](../src/main/resources/static/client.js#L22) file in the editor window.
 
+5. Scroll down to where you see the ```// TODO #2: Add operator``` comment (Line 22)
 
-10. Scroll down to where you see the ```// TODO: Add operator``` comment (Line 22)
-
-11. Move your cursor to the end of the line 35 (to the right of ```break;``` and press **ENTER**.
+6. Move your cursor to the end of the line 35 (to the right of ```break;``` and press **ENTER**.
 
 GitHub Copilot should display ghost text suggesting the code shown in the following screenshot. Press **TAB** to accept the suggestion.
 
 <img width="353" alt="GitHub Copilot suggestions" src="../assets/case-suggestion.png">
 
-12. Press **ENTER** at the end of the line, then accept the next two lines Copilot suggests.
+7. Press **ENTER** at the end of the line, then accept the next two lines Copilot suggests.
 
 Your completed addition should match the following.
 
 <img width="376" alt="GitHub Copilot suggestions" src="../assets/Add-operator-completed.png">
 
-13. Press ```CTRL + ` ``` to open the terminal window in VS Code.
+8. Open the 
+[ArithmeticController.java](../src/main/java/com/copilot/calculator/controller/ArithmeticController.java#L16) file in the editor window.
 
-14. Enter ```./mvnw spring-boot:run``` in the terminal window and press **ENTER** to run the application.
+9. Scroll down to where you see the ```// TODO #3: Add operator``` comment
 
-15. You should test the new button by clicking 3, then the "^" (power) button, then click 2. Click "=" and the result should be 9.
+10. Press **ENTER** at the end of the line that defines the divide function.
 
-16. Close the browser window, return to the Terminal window in Codespaces and press ```CTRL+C``` to terminate the application.
+11. Copy and Paste this line
+```<button class="btn" th:onClick="operationPressed('^')">^</button>```
+
+<img width="353" alt="GitHub Copilot suggestions" src="../assets/power.jpg">
+
+12. Press ```CTRL + ` ``` to open the terminal window in VS Code.
+
+13. Enter ```./mvnw spring-boot:run``` in the terminal window and press **ENTER** to run the application.
+
+14. You should test the new button by clicking 3, then the "^" (power) button, then click 2. Click "=" and the result should be 9.
+
+15. Close the browser window, return to the Terminal window in Codespaces and press ```CTRL+C``` to terminate the application.
 
 **Success**, you have enhanced the calculator application using GitHub Copilot!
