+++ .instructions/3. challenge exercises.md
@@ -16,9 +16,9 @@ Now you've had an opportunity to get started using GitHub Copilot, we have a num
 
 4. Open the [ArithmeticController.java](../src/test/java/com/copilot/calculator/CalculatorApplicationTests.java#83) file.
 
-5. Scroll down to the line with the comment ```TODO: Challenge #1``` (Around line 96)
+5. Scroll down to the line with the comment ```// TODO #4: Generate Test for Power``` 
 
-6. On the line following the comment, add a new comment to provide context to GitHub Copilot on what you want assistance to do. Try adding this comment ```// add tests for subtraction``` and press ```ENTER``` to generate a suggestion.
+6. On the line following the comment, add a new comment to provide context to GitHub Copilot on what you want assistance to do. Try adding this comment ```// add tests for power``` and press ```ENTER``` to generate a suggestion.
 
 7. Accept the suggested line if it looks right by pressing ```TAB``` then ```ENTER```.
 
@@ -33,16 +33,7 @@ Now you've had an opportunity to get started using GitHub Copilot, we have a num
 </details>
 
 <details>
-<summary>Challenge #2 - Adding Unit Tests for the power/exponential function</summary>
-
-### Adding Unit Tests for the power/exponential function
-
-1. See if you can now add additional unit tests for the power/exponential function you created in the core exercise.
-
-</details>
-
-<details>
-<summary>Challenge #3 - Adding a new function</summary>
+<summary>Challenge #2 - Adding a new function</summary>
 
 ### Adding a new function
