--- src/main/java/com/copilot/calculator/IndexController.java
@@ -1,12 +1,0 @@
-package com.copilot.calculator;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-public class IndexController {
-    @RequestMapping("/")
-    public String index() {
-        return "index.html";
-    }
-}
