--- src/main/java/com/copilot/calculator/WebConfig.java
@@ -1,26 +1,0 @@
-package com.copilot.calculator;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.springframework.web.servlet.view.InternalResourceViewResolver;
-
-@Configuration
-@EnableWebMvc
-    public class WebConfig implements WebMvcConfigurer {
-
-        @Override
-        public void addResourceHandlers(ResourceHandlerRegistry registry) {
-            registry.addResourceHandler("/**")
-                    .addResourceLocations("classpath:/static/");
-            
-        }
-        
-        // https://github.com/springdoc/springdoc-openapi/issues/236#issuecomment-567495446
-        @Bean
-        public InternalResourceViewResolver defaultViewResolver() {
-            return new InternalResourceViewResolver();
-        }
-    }
