+++ src/main/java/com/copilot/calculator/controller/IndexController.java
@@ -1,0 +1,13 @@
+package com.copilot.calculator.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+
+@Controller
+public class IndexController {
+    @GetMapping("/")
+    public String index() {
+        return "index";
+    }
+}
