+++ src/test/java/com/copilot/calculator/CalculatorApplicationTests.java
@@ -1,20 +1,20 @@
 package com.copilot.calculator;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import com.copilot.calculator.api.ArithmeticController;
-import static org.assertj.core.api.Assertions.assertThat;
-
+import com.copilot.calculator.controller.ArithmeticController;
 
 @SpringBootTest
 class CalculatorApplicationTests {
 
 	@Autowired
-    private ArithmeticController arithmeticController;
+	private ArithmeticController arithmeticController;
+
 	@Test
 	void contextLoads() {
 	}
@@ -22,30 +22,31 @@ class CalculatorApplicationTests {
 	// Validation Test
 	@Test
 	public void testInvalidOperation() throws Exception {
-		//rejects missing operation, expects IllegalArgumentException("Invalid operation: "
+		// rejects missing operation, expects IllegalArgumentException("Invalid
+		// operation: "
 		var exception = assertThrows(IllegalArgumentException.class, () -> {
-        	arithmeticController.calculate("", "1", "2");
+			arithmeticController.calculate("", "1", "2");
 		});
 		assertThat(exception.getMessage()).isEqualTo("Invalid operation: ");
 
-		//rejects invalid operation
+		// rejects invalid operation
 		exception = assertThrows(IllegalArgumentException.class, () -> {
 			arithmeticController.calculate("invalid", "1", "2");
 		});
 		assertThat(exception.getMessage()).isEqualTo("Invalid operation: invalid");
 
-		//rejects missing operand1
+		// rejects missing operand1
 		exception = assertThrows(IllegalArgumentException.class, () -> {
 			arithmeticController.calculate("add", "", "2");
 		});
 		assertThat(exception.getMessage()).isEqualTo("Operand cannot be null or empty");
 
-		//rejects missing operand2
+		// rejects missing operand2
 		exception = assertThrows(IllegalArgumentException.class, () -> {
 			arithmeticController.calculate("add", "1", "");
 		});
 		assertThat(exception.getMessage()).isEqualTo("Operand cannot be null or empty");
-		
+
 	}
 
 	// Test Addition
@@ -80,6 +81,5 @@ class CalculatorApplicationTests {
 		assertThat(arithmeticController.calculate("divide", "-1", "-2")).isEqualTo("{\"result\": 0.50}");
 	}
 
-	//TODO: Challenge #1
-	
+	// TODO #4: Challenge
 }
